# Conflicts:
#	.github/workflows/frontend.yml
#	.github/workflows/release.yml
#	CHANGELOG/README.md
#	scripts/bk-ci-special-upgrade.sh
#	scripts/bkce-set-env03.sh
#	scripts/deploy-codecc/bk-codecc-install.sh
#	scripts/deploy-codecc/bkce-set-env03-codecc.sh
#	src/agent/src/pkg/logs/logs.go
#	src/backend/ci/build.gradle.kts
#	src/backend/ci/buildSrc/src/main/kotlin/constants/Versions.kt
#	src/backend/ci/core/common/common-api/build.gradle.kts
#	src/backend/ci/core/common/common-api/src/main/kotlin/com/tencent/devops/common/api/util/JsonUtil.kt
#	src/backend/ci/core/misc/model-image/build.gradle.kts
#	src/backend/ci/core/openapi/api-openapi/build.gradle.kts
#	src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/PipelineBuildDetailService.kt
#	src/backend/ci/core/process/biz-base/src/main/kotlin/com/tencent/devops/process/engine/service/detail/TaskBuildDetailService.kt
#	src/backend/ci/core/quality/biz-quality/src/main/kotlin/com/tencent/devops/quality/service/v2/QualityIndicatorService.kt
#	src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/service/RepositoryAuthService.kt
#	src/backend/ci/core/repository/biz-repository/src/main/kotlin/com/tencent/devops/repository/service/RepositoryService.kt
#	src/backend/ci/core/worker/worker-common/src/main/kotlin/com/tencent/devops/worker/WorkRunner.kt
#	src/backend/codecc/build.gradle.kts
#	src/backend/codecc/buildSrc/src/main/kotlin/Versions.kt
#	src/backend/codecc/core/codeccjob/biz-codeccjob/src/main/java/com/tencent/bk/codecc/codeccjob/consumer/WebsocketConsumer.java
#	src/backend/codecc/core/common/common-client/common-client-base/src/main/kotlin/com/tencent/devops/common/client/Client.kt
#	src/backend/codecc/core/common/common-client/common-client-consul/src/main/kotlin/com/tencent/devops/common/client/ConsulClientAutoConfiguration.kt
#	src/backend/codecc/core/common/common-client/common-client-k8s/src/main/kotlin/com/tencent/devops/common/client/KubernetesClientAutoConfiguration.kt
#	src/backend/codecc/core/common/common-util/build.gradle.kts
#	src/backend/codecc/core/common/common-web/src/main/kotlin/com/tencent/devops/common/web/mq/Constants.kt
#	src/backend/codecc/core/defect/biz-defect-migration/src/main/kotlin/com/tencent/bk/codecc/defect/service/impl/PipelineScmServiceImpl.kt
#	src/backend/codecc/core/defect/biz-defect/build.gradle.kts
#	src/backend/codecc/core/defect/biz-defect/src/main/java/com/tencent/bk/codecc/defect/service/AbstractQueryWarningBizService.java
#	src/backend/codecc/core/defect/biz-defect/src/main/java/com/tencent/bk/codecc/defect/service/impl/CCNQueryWarningBizServiceImpl.java
#	src/backend/codecc/core/defect/biz-defect/src/main/java/com/tencent/bk/codecc/defect/service/impl/CommonQueryWarningBizServiceImpl.java
#	src/backend/codecc/core/defect/biz-defect/src/main/java/com/tencent/bk/codecc/defect/service/impl/DUPCQueryWarningBizServiceImpl.java
#	src/backend/codecc/core/defect/biz-defect/src/main/kotlin/com/tencent/bk/codecc/defect/service/PipelineScmService.kt
#	src/backend/codecc/core/defect/biz-defect/src/main/kotlin/com/tencent/bk/codecc/defect/service/impl/TaskLogOverviewServiceImpl.kt
#	src/backend/codecc/core/defect/model-defect/build.gradle.kts
#	src/backend/codecc/core/quartz/biz-quartz/src/main/kotlin/com/tencent/bk/codecc/quartz/service/impl/ShardingRouterServiceImpl.kt
#	src/backend/codecc/core/quartz/biz-quartz/src/main/kotlin/com/tencent/bk/codecc/quartz/strategy/sharding/impl/AscendShardingStrategy.kt
#	src/backend/codecc/core/quartz/sdk-quartz/src/main/kotlin/com/tencent/bk/codecc/quartz/job/ActiveGongfengProjScheduleTask.kt
#	src/backend/codecc/core/task/biz-task-migration/src/main/kotlin/com/tencent/bk/codecc/task/service/impl/PipelineServiceImpl.kt
#	src/backend/codecc/settings.gradle.kts
#	src/backend/turbo/build.gradle.kts
#	src/backend/turbo/common-turbo/common-turbo-client/src/main/kotlin/com/tencent/devops/common/client/JerseyFeignClientFactoryBean.kt
#	src/backend/turbo/support-files/templates/#etc#ci#application-turbo.yml
#	src/frontend/devops-artifactory/package.json
#	src/frontend/devops-atomstore/package.json
#	src/frontend/devops-codelib/package.json
#	src/frontend/devops-environment/package.json
#	src/frontend/devops-environment/src/components/devops/environment/third-construct-dialog.vue
#	src/frontend/devops-nav/package.json
#	src/frontend/devops-nav/src/components/Header/NavMenu.vue
#	src/frontend/devops-nav/src/components/Header/index.vue
#	src/frontend/devops-nav/src/components/NavBox/index.vue
#	src/frontend/devops-nav/src/index.html
#	src/frontend/devops-nav/src/views/Home.vue
#	src/frontend/devops-nav/src/views/ProjectManage.vue
#	src/frontend/devops-pipeline/package.json
#	src/frontend/devops-pipeline/src/components/AtomSelector/atomCard.vue
#	src/frontend/devops-pipeline/src/components/AtomSelector/index.vue
#	src/frontend/devops-pipeline/src/components/pipelineSetting/BaseInfo.vue
#	src/frontend/devops-pipeline/src/store/modules/atom/actions.js
#	src/frontend/devops-pipeline/src/store/modules/atom/constants.js
#	src/frontend/devops-pipeline/src/store/modules/atom/getters.js
#	src/frontend/devops-pipeline/src/store/modules/atom/mutations.js
#	src/frontend/devops-pipeline/src/views/list/atomManage.vue
#	src/frontend/devops-pipeline/src/views/list/newlist.vue
#	src/frontend/devops-quality/package.json
#	src/frontend/devops-quality/src/views/overview.vue
#	src/frontend/devops-ticket/package.json
#	src/frontend/devops-turbo/package.json
#	src/frontend/devops-turbo/src/api/index.js
#	src/frontend/locale/nav/en-US.json
#	src/frontend/locale/nav/zh-CN.json
#	src/frontend/locale/pipeline/en-US.json
#	src/frontend/locale/pipeline/zh-CN.json
#	src/gateway/core/proxy.set.header.common.conf
#	support-files/codecc/kubernetes/images/gateway/conf/lua/auth/auth_user-bk_login.lua
#	support-files/codecc/kubernetes/images/gateway/conf/lua/auth/auth_user-bk_login_v3.lua
#	support-files/sql/5001_ci_quality-init_dml_mysql.sql
#	support-files/sql/5001_ci_store-init_dml_mysql.sql

+++ src/backend/ci/core/misc/biz-misc/src/main/kotlin/com/tencent/devops/misc/service/process/ProcessDataClearService.kt
@@ -27,51 +27,76 @@
 
 package com.tencent.devops.misc.service.process
 
+import com.tencent.devops.common.db.pojo.ARCHIVE_SHARDING_DSL_CONTEXT
+import com.tencent.devops.common.db.utils.JooqUtils
+import com.tencent.devops.common.redis.RedisOperation
 import com.tencent.devops.misc.dao.process.ProcessDao
 import com.tencent.devops.misc.dao.process.ProcessDataClearDao
+import com.tencent.devops.misc.lock.PipelineVersionLock
 import org.jooq.DSLContext
 import org.jooq.impl.DSL
 import org.springframework.beans.factory.annotation.Autowired
+import org.springframework.beans.factory.annotation.Qualifier
 import org.springframework.stereotype.Service
 
 @Service
 class ProcessDataClearService @Autowired constructor(
     private val dslContext: DSLContext,
+    @Qualifier(ARCHIVE_SHARDING_DSL_CONTEXT)
+    private val archiveShardingDslContext: DSLContext,
     private val processDao: ProcessDao,
-    private val processDataClearDao: ProcessDataClearDao
+    private val processDataClearDao: ProcessDataClearDao,
+    private val redisOperation: RedisOperation,
+    private val processRelatedPlatformDataClearService: ProcessRelatedPlatformDataClearService
 ) {
 
     /**
      * 清除流水线数据
      * @param projectId 项目ID
      * @param pipelineId 流水线ID
+     * @param archiveFlag 归档标识
      */
     fun clearPipelineData(
         projectId: String,
-        pipelineId: String
+        pipelineId: String,
+        archiveFlag: Boolean? = null
     ) {
-        dslContext.transaction { t ->
+        val finalDslContext = generateFinalDslContext(archiveFlag)
+        finalDslContext.transaction { t ->
             val context = DSL.using(t)
             processDataClearDao.deletePipelineLabelByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineModelTaskByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineRemoteAuthByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineResourceByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineResourceVersionByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineSettingByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineSettingVersionByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineTimerByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineWebhookByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deleteTemplatePipelineByPipelineId(context, projectId, pipelineId)
             processDataClearDao.deletePipelineBuildSummaryByPipelineId(context, projectId, pipelineId)
-            processDataClearDao.deletePipelineTemplateAcrossInfo(context, projectId, pipelineId)
-            // 添加删除记录，插入要实现幂等
-            processDao.addPipelineDataClear(
-                dslContext = context,
-                projectId = projectId,
-                pipelineId = pipelineId
-            )
+            if (archiveFlag != true) {
+                processDataClearDao.deletePipelineModelTaskByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineRemoteAuthByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineSettingByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineSettingVersionByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineTimerByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineWebhookByPipelineId(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineTemplateAcrossInfo(context, projectId, pipelineId)
+                processDataClearDao.deletePipelineViewGroup(context, projectId, pipelineId)
+                // 添加删除记录，插入要实现幂等
+                processDao.addPipelineDataClear(
+                    dslContext = context,
+                    projectId = projectId,
+                    pipelineId = pipelineId
+                )
+            }
             processDataClearDao.deletePipelineInfoByPipelineId(context, projectId, pipelineId)
         }
+        processRelatedPlatformDataClearService.cleanBuildData(projectId, pipelineId)
+    }
+
+    private fun generateFinalDslContext(archiveFlag: Boolean?): DSLContext {
+        val finalDslContext = if (archiveFlag == true) {
+            archiveShardingDslContext
+        } else {
+            dslContext
+        }
+        return finalDslContext
     }
 
     /**
@@ -94,14 +119,26 @@ class ProcessDataClearService @Autowired constructor(
      * @param projectId 项目ID
      * @param pipelineId 流水线ID
      * @param buildId 构建ID
+     * @param archiveFlag 归档标识
      */
     fun clearOtherBuildData(
         projectId: String,
         pipelineId: String,
-        buildId: String
+        buildId: String,
+        archiveFlag: Boolean? = null
     ) {
-        dslContext.transaction { t ->
+        val finalDslContext = generateFinalDslContext(archiveFlag)
+        finalDslContext.transaction { t ->
             val context = DSL.using(t)
+            processDataClearDao.deleteBuildRecordPipelineByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordStageByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordContainerByBuildId(context, projectId, buildId)
+            processDataClearDao.deleteBuildRecordTaskByBuildId(context, projectId, buildId)
+            if (archiveFlag == true) {
+                processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+                // 归档库构建数据清理无需执行方法后的逻辑
+                return@transaction
+            }
             processDataClearDao.deleteBuildDetailByBuildId(context, projectId, buildId)
             processDataClearDao.deleteReportByBuildId(
                 dslContext = context,
@@ -109,7 +146,10 @@ class ProcessDataClearService @Autowired constructor(
                 pipelineId = pipelineId,
                 buildId = buildId
             )
-            processDataClearDao.deletePipelineBuildTemplateAcrossInfo(context, projectId, buildId)
+            JooqUtils.retryWhenDeadLock {
+                processDataClearDao.deletePipelineBuildTemplateAcrossInfo(context, projectId, pipelineId, buildId)
+            }
+            processDataClearDao.deleteBuildWebhookParameter(context, projectId, buildId)
             // 添加删除记录，插入要实现幂等
             processDao.addBuildHisDataClear(
                 dslContext = context,
@@ -117,7 +157,71 @@ class ProcessDataClearService @Autowired constructor(
                 pipelineId = pipelineId,
                 buildId = buildId
             )
-            processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+            val version = processDao.getPipelineVersionByBuildId(
+                dslContext = context,
+                projectId = projectId,
+                pipelineId = pipelineId,
+                buildId = buildId
+            )
+            val pipelineVersionLock = PipelineVersionLock(redisOperation, pipelineId, version)
+            try {
+                pipelineVersionLock.lock()
+                val deleteResult = processDataClearDao.deleteBuildHistoryByBuildId(context, projectId, buildId)
+                if (deleteResult == 0) {
+                    // 如果删除的记录数为0则无需执行后面的逻辑
+                    return@transaction
+                }
+                // 查询流水线版本记录
+                val pipelineVersionInfo = processDao.getPipelineVersionSimple(
+                    dslContext = context,
+                    projectId = projectId,
+                    pipelineId = pipelineId,
+                    version = version
+                )
+                var referCount = pipelineVersionInfo?.referCount
+                referCount = if (referCount == null || referCount < 0) {
+                    // 兼容老数据缺少关联构建记录的情况，全量统计关联数据数量
+                    processDao.countBuildNumByVersion(
+                        dslContext = context,
+                        projectId = projectId,
+                        pipelineId = pipelineId,
+                        version = version
+                    )
+                } else {
+                    referCount - 1
+                }
+                val referFlag = referCount > 0
+                // 更新流水线版本关联构建记录信息
+                processDao.updatePipelineVersionReferInfo(
+                    dslContext = context,
+                    projectId = projectId,
+                    pipelineId = pipelineId,
+                    version = version,
+                    referCount = referCount,
+                    referFlag = referFlag
+                )
+            } finally {
+                pipelineVersionLock.unlock()
+            }
         }
+    }
+
+    /**
+     * 清除流水线被跳过的任务数据
+     * @param projectId 项目ID
+     * @param buildId 构建ID
+     * @param archiveFlag 归档标识
+     */
+    fun clearSkipRecordTaskData(
+        projectId: String,
+        buildId: String,
+        archiveFlag: Boolean? = null
+    ) {
+        processDataClearDao.deleteBuildRecordTaskByBuildId(
+            dslContext = generateFinalDslContext(archiveFlag),
+            projectId = projectId,
+            buildId = buildId,
+            skipTaskDeleteFlag = true
+        )
     }
 }
